# pkg/api/docs/swagger.yaml
basePath: /
definitions:
...
  CreateTargetDTO:
    properties:
      id:
        type: string
      name:
        type: string
      targetConfigName:
        type: string
    required:
    - id
    - name
    - targetConfigName
    type: object
...
  Target:
    properties:
      default:
        type: boolean
      id:
        type: string
      name:
        type: string
      options:
        description: JSON encoded map of options
        type: string
      providerInfo:
        $ref: '#/definitions/TargetProviderInfo'
    required:
    - default
    - id
    - name
    - options
    - providerInfo
    type: object
...
  TargetDTO:
    properties:
      default:
        type: boolean
      id:
        type: string
      info:
        $ref: '#/definitions/TargetInfo'
      name:
        type: string
      options:
        description: JSON encoded map of options
        type: string
      providerInfo:
        $ref: '#/definitions/TargetProviderInfo'
      workspaceCount:
        type: integer
    required:
    - default
    - id
    - name
    - options
    - providerInfo
    - workspaceCount
    type: object
  TargetInfo:
    properties:
      name:
        type: string
      providerMetadata:
        type: string
    required:
    - name
    type: object
  TargetProviderInfo:
    properties:
      label:
        type: string
      name:
        type: string
      version:
        type: string
    required:
    - name
    - version
    type: object
...
host: localhost:3986
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: v0.0.0-dev
paths:
...
  /target:
    get:
      description: List targets
      operationId: ListTargets
      parameters:
      - description: Verbose
        in: query
        name: verbose
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TargetDTO'
            type: array
      summary: List targets
      tags:
      - target
    post:
      description: Create a target
      operationId: CreateTarget
      parameters:
      - description: Create target
        in: body
        name: target
        required: true
        schema:
          $ref: '#/definitions/CreateTargetDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Target'
      summary: Create a target
      tags:
      - target
...
  /target-config/{configName}:
    delete:
      description: Remove a target config
      operationId: RemoveTargetConfig
      parameters:
      - description: Target Config name
        in: path
        name: configName
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Remove a target config
      tags:
      - target-config
  /target/{targetId}:
    delete:
      description: Remove target
      operationId: RemoveTarget
      parameters:
      - description: Target ID
        in: path
        name: targetId
        required: true
        type: string
      - description: Force
        in: query
        name: force
        type: boolean
      responses:
        "200":
          description: OK
      summary: Remove target
      tags:
      - target
    get:
      description: Get target info
      operationId: GetTarget
      parameters:
      - description: Target ID or Name
        in: path
        name: targetId
        required: true
        type: string
      - description: Verbose
        in: query
        name: verbose
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TargetDTO'
      summary: Get target info
      tags:
      - target
...
schemes:
- http
security:
- Bearer: []
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"